--- composer.json
@@ -1,39 +1,0 @@
-{
-    "name": "vitodeploy/laravel-octane-plugin",
-    "description": "Laravel octane plugin for VitoDeploy",
-    "type": "library",
-    "license": "MIT",
-    "version": "1.0.0",
-    "authors": [
-        {
-            "name": "Saeed Vaziry",
-            "email": "mr.saeedvaziry@gmail.com"
-        }
-    ],
-    "scripts": {
-        "test": "vendor/bin/phpunit"
-    },
-    "require": {
-        "php": "^8.4",
-        "illuminate/support": "^12.0",
-        "illuminate/process": "^12.0",
-        "illuminate/console": "^12.0"
-    },
-    "require-dev": {
-        "laravel/pint": "^1.10"
-    },
-    "autoload": {
-        "psr-4": {
-            "VitoDeploy\\LaravelOctanePlugin\\": "src"
-        }
-    },
-    "extra": {
-        "laravel": {
-            "providers": [
-                "VitoDeploy\\LaravelOctanePlugin\\LaravelOctanePluginServiceProvider"
-            ]
-        }
-    },
-    "minimum-stability": "stable",
-    "prefer-stable": true
-}
